+++ src/translations/lang_bg_BG.ts
@@ -756,7 +756,7 @@ Check the permissions of the destination file and directory.</source>
 <context>
     <name>SysTray</name>
     <message>
-        <location filename="../app/systray.cpp" line="71"/>
+        <location filename="../app/systray.cpp" line="70"/>
         <source>Quit</source>
         <translation>Затваряне</translation>
     </message>
@@ -955,6 +955,11 @@ Check the permissions of the destination file and directory.</source>
         <location filename="../core/systeminfoui.cpp" line="68"/>
         <source>l2cache</source>
         <translation>L2 кеш</translation>
+    </message>
+    <message>
+        <location filename="../core/systeminfoui.cpp" line="69"/>
+        <source>uniqueid</source>
+        <translation>Уникален идентификатор</translation>
     </message>
     <message>
         <location filename="../core/systeminfoui.cpp" line="57"/>