+++ src/translations/lang_en_EN.ts
@@ -756,7 +756,7 @@ Check the permissions of the destination file and directory.</translation>
 <context>
     <name>SysTray</name>
     <message>
-        <location filename="../app/systray.cpp" line="71"/>
+        <location filename="../app/systray.cpp" line="70"/>
         <source>Quit</source>
         <translation>Quit</translation>
     </message>
@@ -955,6 +955,11 @@ Check the permissions of the destination file and directory.</translation>
         <location filename="../core/systeminfoui.cpp" line="68"/>
         <source>l2cache</source>
         <translation>L2 cache</translation>
+    </message>
+    <message>
+        <location filename="../core/systeminfoui.cpp" line="69"/>
+        <source>uniqueid</source>
+        <translation>Unique ID</translation>
     </message>
     <message>
         <location filename="../core/systeminfoui.cpp" line="57"/>