+++ src/translations/lang_ru_RU.ts
@@ -757,7 +757,7 @@ Check the permissions of the destination file and directory.</source>
 <context>
     <name>SysTray</name>
     <message>
-        <location filename="../app/systray.cpp" line="71"/>
+        <location filename="../app/systray.cpp" line="70"/>
         <source>Quit</source>
         <translation>Завершить</translation>
     </message>
@@ -956,6 +956,11 @@ Check the permissions of the destination file and directory.</source>
         <location filename="../core/systeminfoui.cpp" line="68"/>
         <source>l2cache</source>
         <translation>Кэш L2</translation>
+    </message>
+    <message>
+        <location filename="../core/systeminfoui.cpp" line="69"/>
+        <source>uniqueid</source>
+        <translation>Уникальный идентификатор</translation>
     </message>
     <message>
         <location filename="../core/systeminfoui.cpp" line="57"/>